Digipeater Cleanup and move
